pyproject.toml
[tool.poetry]
name = "marpla"
version = "0.1.0"
description = ""
authors = ["Egor <wzxv_@outlook.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
psycopg2-binary = "^2.9.9"
confluent-kafka = "^2.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
